function/requirements.txt
azure-functions
psycopg2
sendgrid